requirements.txt
Django==5.1.*
django-cors-headers==4.5.*
django-filter==24.3
django-rest-swagger==2.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-user-agents==0.4.0
drf-yasg==1.21.*
gunicorn==23.0.*
pillow==10.4.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
django-active-link==0.2.*
